factored drag and zoom function into a separate VisualizerContainer
