+++ client/src/components/Array1D/Array1D.css
@@ -24,43 +24,11 @@
     background-color: #b3beff;
 }
 
-.table-container {
-    display: flex;
-    margin: 0px auto;
-    width: auto;
-    height: 200px;
-    outline: 1px dotted black;
-    overflow: hidden;
-}
-
-.reset-position-button > * {
-    font-size: 12px;
-    padding: 2px 4px;
-    border: none;
-    float: left;
-    position: relative;
-}
-
-.reset-position-button .fa {
-    margin: 0;
-}
-
-.wrapper {
-    margin: auto;
-}
-
 .elements {
     table-layout: fixed;
     margin: auto;
     border-collapse: separate;
     border-spacing: 10px 0;
-}
-.elements:hover {
-    cursor:grab
-}
-
-.elements:active {
-    cursor: grabbing;
 }
 
 .elements td {
